videoflix/requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
click==8.1.7
colorama==0.4.6
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.5
django-debug-toolbar==4.2.0
django-environ==0.11.2
django-import-export==3.3.1
django-redis==5.3.0
django-rq==2.8.1
djangorestframework==3.14.0
et-xmlfile==1.1.0
gunicorn==21.2.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.2
psycopg2-binary==2.9.8
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
rq==1.15.1
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.8.0
tzdata==2023.3
xlrd==2.0.1
xlwt==1.3.0
